Database sample cleanup (#545)
