+++ app/src/main/res/values/strings.xml
@@ -64,10 +64,10 @@
     <string name="choose_image">Choose Image</string>
     <string name="accessibility_downloaded_image">Downloaded image</string>
     <string name="drive_file">Drive File</string>
+    <string name="allow_new_email_acccount">Allow account creation if email does not exist.</string>
 
-    <!-- strings for Auth UI demo activities -->
+    <!-- strings for database demo activities -->
+    <string name="start_chatting">No messages. Start chatting at the bottom!</string>
     <string name="signed_in">Signed In</string>
     <string name="sign_in_failed">Sign In Failed</string>
-    <string name="allow_new_email_acccount">Allow account creation if email does not exist.</string>
-    <string name="start_chatting">No messages. Start chatting at the bottom!</string>
 </resources>
